Update README.md
Each folder contains an explanation of the dynamical system it is named after, code to picture/animate it, a subfolder names "media" that has outputs, and instructions on using it.

Occasionally a project will take me many revisions to get right. In this case, a subfolder named "old" will have older versions of the main file. The code will generally run, but it is slow and achieves the goal in a different way.

The pictures/animations in the media folder have often been edited from their programmed output.

Follow me on Twitter @tyler_russell52 or email me at tyler.russell1@aol.com for questions.